Move "Header" trait for toml::Value to store module
