+++ lib/core/libimagstore/src/toml_ext.rs
@@ -362,97 +362,6 @@ impl Extract for Value {
 
 pub type EntryResult<T> = RResult<T, ParserError>;
 
-/// Extension trait for top-level toml::Value::Table, will only yield correct results on the
-/// top-level Value::Table, but not on intermediate tables.
-pub trait Header {
-    fn verify(&self) -> Result<()>;
-    fn parse(s: &str) -> EntryResult<Value>;
-    fn default_header() -> Value;
-}
-
-impl Header for Value {
-
-    fn verify(&self) -> Result<()> {
-        match *self {
-            Value::Table(ref t) => verify_header(&t),
-            _ => Err(SE::new(SEK::HeaderTypeFailure, None)),
-        }
-    }
-
-    fn parse(s: &str) -> EntryResult<Value> {
-        use toml::de::from_str;
-
-        from_str(s)
-            .map_err(|_| ParserErrorKind::TOMLParserErrors.into())
-            .and_then(verify_header_consistency)
-            .map(Value::Table)
-    }
-
-    fn default_header() -> Value {
-        let mut m = BTreeMap::new();
-
-        m.insert(String::from("imag"), {
-            let mut imag_map = BTreeMap::<String, Value>::new();
-
-            imag_map.insert(String::from("version"), Value::String(String::from(version!())));
-            imag_map.insert(String::from("links"), Value::Array(vec![]));
-
-            Value::Table(imag_map)
-        });
-
-        Value::Table(m)
-    }
-
-}
-
-pub fn verify_header_consistency(t: Table) -> EntryResult<Table> {
-    verify_header(&t)
-        .map_err(Box::new)
-        .map_err(|e| ParserErrorKind::HeaderInconsistency.into_error_with_cause(e))
-        .map(|_| t)
-}
-
-fn verify_header(t: &Table) -> Result<()> {
-    if !has_main_section(t) {
-        Err(SE::from(ParserErrorKind::MissingMainSection.into_error()))
-    } else if !has_imag_version_in_main_section(t) {
-        Err(SE::from(ParserErrorKind::MissingVersionInfo.into_error()))
-    } else if !has_only_tables(t) {
-        debug!("Could not verify that it only has tables in its base table");
-        Err(SE::from(ParserErrorKind::NonTableInBaseTable.into_error()))
-    } else {
-        Ok(())
-    }
-}
-
-fn has_only_tables(t: &Table) -> bool {
-    debug!("Verifying that table has only tables");
-    t.iter().all(|(_, x)| is_match!(*x, Value::Table(_)))
-}
-
-pub fn has_main_section(t: &Table) -> bool {
-    t.contains_key("imag") && is_match!(t.get("imag"), Some(&Value::Table(_)))
-}
-
-pub fn has_imag_version_in_main_section(t: &Table) -> bool {
-    use semver::Version;
-
-    match *t.get("imag").unwrap() {
-        Value::Table(ref sec) => {
-            sec.get("version")
-                .and_then(|v| {
-                    match *v {
-                        Value::String(ref s) => Some(Version::parse(&s[..]).is_ok()),
-                        _                    => Some(false),
-                    }
-                })
-            .unwrap_or(false)
-        }
-        _ => false,
-    }
-}
-
-
 #[cfg(test)]
 mod test {
     extern crate env_logger;
